rename: ClientIO.type -> ClientIOTypes
